Display nominated, recent maps in nomination menu
We want to display all maps that are in the rotation, regardless
of whether they have been recently played, nominated, are
currently being played or are otherwise temporarily unavailable.

This prevents people from coming to the Uncletopia Discord to
cry about how their favourite map has been removed (no, nerd,
it was just not played recently).